#161 Moved parse_date outside of gtfs_stats and removed compute_route_stats_time_cutoff
